requirements.txt
torch
transformers
pillow
numpy
faiss-cpu
requests
python-telegram-bot
datasets
